refactor(realm): extract exec bridge from js-realm-shared
